Correct git --orphan
The `git checkout --orphan` command will make all the files
automatically tracked. This means that all the files from the
template branch will be added. Because the `rsync` step comes
after, we are unable to remove files that would otherwise be
undesired.

I explored adding `--delete-excluded` to the **rsync** command,
but there is a catch to this, the `.git` folder is also deleted
since it's being excluded. So this proves to be a bad solution.